\name{CreateReviewObject}
\alias{CreateReviewObject}
\title{
A function to summarize multiple ABSOLUTE runs
}
\description{
This function will take in a collection of ABSOLUTE results and create
summary tables & plots.
}
\usage{
CreateReviewObject(obj.name, absolute.files, indv.results.dir, copy_num_type, 
                   plot.modes=TRUE, verbose=FALSE)
}
\arguments{
  \item{obj.name}{
    A descriptive name for this collection
  } 
  \item{absolute.files}{
   One or more RData files from \code{RunAbsolute} calls
 }
  \item{indv.results.dir}{
    The directory to store the output files generated by this call
  }
  \item{copy_num_type}{
    The type of copy number to be handled. Either \code{total} or \code{allelic}
  }
  \item{plot.modes}{
    If \code{TRUE} will plot the purity/ploidy modes
  }
  \item{verbose}{
    Prints more verbose output when \code{TRUE}
  }
}
\value{
  \code{CreateReviewObject} simply returns \code{TRUE} on success, but will write a
  combination of files as a side effect, all within \code{results.dir}.
  
  The \code{OBJNAME_PP-modes_data.RData} file contains an object \code{segobj_list}
  which has an element for every sample which corresponds to that sample's
  \code{seg.dat}.

  The \code{OBJNAME_PP-modes_plotes.pdf} provides summary images for all of the
  samples processed.

  The \code{OBJNAME_PP-calls_tab.txt} file contains a summary table of all of
  the samples. 

  The \code{sample.plots} directory contains the per-sample PDF plots from the
  individuals \code{RunAbsolute} runs.
}
\author{
Scott L. Carter
}
\seealso{
\code{\link{RunAbsolute}}
}
\examples{
}
\keyword{ hplot }
\keyword{ utilities }
